@@ -5,24 +5,25 @@
     "license": "MIT",
     "require": {
         "php": "^5.5 || ^7.0",
-        "sulu/sulu": "^1.5.3 || dev-develop",
+        "sulu/sulu": "^1.6.8",
         "ongr/elasticsearch-bundle": "^1.2.9 || ~5.0",
-        "ongr/elasticsearch-dsl": "^2.2.2 || ~5.0"
+        "ongr/elasticsearch-dsl": "^2.2.2 || ~5.0",
+        "jackalope/jackalope": "^1.2.8 || >=1.3.3",
+        "jms/serializer-bundle": "^1.1 || ^2.0"
     },
     "require-dev": {
-        "zendframework/zend-stdlib": "2.3.1 as 2.0.0rc5",
-        "zendframework/zendsearch": "2.*@dev",
-        "symfony/monolog-bundle": "2.4.*",
-        "doctrine/doctrine-bundle": "1.4.*",
+        "massive/build-bundle": "~0.3.0",
+        "zendframework/zend-stdlib": "~2.3",
+        "zendframework/zendsearch": "@dev",
+        "symfony/monolog-bundle": "^2.8.7 || ^3.0",
         "jackalope/jackalope-doctrine-dbal": "^1.2.5",
-        "jackalope/jackalope-jackrabbit": "1.2.*",
+        "jackalope/jackalope-jackrabbit": "^1.2",
         "phpunit/phpunit": "^4.8 || ^5.0",
-
-        "sulu/document-manager": "@dev",
-        "massive/search-bundle": "@dev",
-
-        "php-task/task-bundle": "@dev",
-        "php-task/php-task": "@dev"
+        "sulu/automation-bundle": "^1.2",
+        "php-task/task-bundle": "^1.2"
+    },
+    "suggest": {
+        "sulu/automation-bundle": "Allows to outsource long-running route update processes."
     },
     "keywords": [
         "articles"
