@@ -11,7 +11,6 @@
 
 namespace Sulu\Bundle\ArticleBundle\Content;
 
-use ONGR\ElasticsearchBundle\Result\DocumentIterator;
 use ONGR\ElasticsearchBundle\Service\Manager;
 use ONGR\ElasticsearchDSL\Query\Compound\BoolQuery;
 use ONGR\ElasticsearchDSL\Query\MatchAllQuery;
@@ -22,53 +21,80 @@ use ProxyManager\Factory\LazyLoadingValueHolderFactory;
 use ProxyManager\Proxy\LazyLoadingInterface;
 use Sulu\Bundle\ArticleBundle\Document\ArticleDocument;
 use Sulu\Bundle\ArticleBundle\Document\ArticleViewDocumentInterface;
+use Sulu\Bundle\WebsiteBundle\ReferenceStore\ReferenceStoreInterface;
 use Sulu\Component\Content\Compat\PropertyParameter;
 use Sulu\Component\DocumentManager\DocumentManagerInterface;
 use Sulu\Component\SmartContent\Configuration\Builder;
+use Sulu\Component\SmartContent\Configuration\BuilderInterface;
+use Sulu\Component\SmartContent\DataProviderAliasInterface;
 use Sulu\Component\SmartContent\DataProviderInterface;
 use Sulu\Component\SmartContent\DataProviderResult;
 
 /**
  * Introduces articles in smart-content.
  */
-class ArticleDataProvider implements DataProviderInterface
+class ArticleDataProvider implements DataProviderInterface, DataProviderAliasInterface
 {
     /**
      * @var Manager
      */
-    private $searchManager;
+    protected $searchManager;
 
     /**
      * @var DocumentManagerInterface
      */
-    private $documentManager;
+    protected $documentManager;
 
     /**
      * @var LazyLoadingValueHolderFactory
      */
-    private $proxyFactory;
+    protected $proxyFactory;
+
+    /**
+     * @var ReferenceStoreInterface
+     */
+    private $referenceStore;
+
+    /**
+     * @var ArticleResourceItemFactory
+     */
+    protected $articleResourceItemFactory;
 
     /**
      * @var string
      */
-    private $articleDocumentClass;
+    protected $articleDocumentClass;
+
+    /**
+     * @var int
+     */
+    protected $defaultLimit;
 
     /**
      * @param Manager $searchManager
      * @param DocumentManagerInterface $documentManager
      * @param LazyLoadingValueHolderFactory $proxyFactory
-     * @param $articleDocumentClass
+     * @param ReferenceStoreInterface $referenceStore
+     * @param ArticleResourceItemFactory $articleResourceItemFactory
+     * @param string $articleDocumentClass
+     * @param int $defaultLimit
      */
     public function __construct(
         Manager $searchManager,
         DocumentManagerInterface $documentManager,
         LazyLoadingValueHolderFactory $proxyFactory,
-        $articleDocumentClass
+        ReferenceStoreInterface $referenceStore,
+        ArticleResourceItemFactory $articleResourceItemFactory,
+        $articleDocumentClass,
+        $defaultLimit
     ) {
         $this->searchManager = $searchManager;
         $this->documentManager = $documentManager;
         $this->proxyFactory = $proxyFactory;
+        $this->referenceStore = $referenceStore;
+        $this->articleResourceItemFactory = $articleResourceItemFactory;
         $this->articleDocumentClass = $articleDocumentClass;
+        $this->defaultLimit = $defaultLimit;
     }
 
     /**
@@ -76,13 +102,23 @@ class ArticleDataProvider implements DataProviderInterface
      */
     public function getConfiguration()
     {
+        return $this->getConfigurationBuilder()->getConfiguration();
+    }
+
+    /**
+     * Create new configuration-builder.
+     *
+     * @return BuilderInterface
+     */
+    protected function getConfigurationBuilder()
+    {
         return Builder::create()
             ->enableTags()
             ->enableCategories()
             ->enableLimit()
             ->enablePagination()
             ->enablePresentAs()
-            ->setDeepLink('article/{locale}/edit:{id}/details')
+            ->setDeepLink('articles/{locale}/edit:{id}/details')
             ->enableSorting(
                 [
                     ['column' => 'published', 'title' => 'sulu_article.smart-content.published'],
@@ -91,8 +127,7 @@ class ArticleDataProvider implements DataProviderInterface
                     ['column' => 'title', 'title' => 'sulu_article.smart-content.title'],
                     ['column' => 'author_full_name', 'title' => 'sulu_article.smart-content.author-full-name'],
                 ]
-            )
-            ->getConfiguration();
+            );
     }
 
     /**
@@ -115,18 +150,17 @@ class ArticleDataProvider implements DataProviderInterface
         $pageSize = null
     ) {
         $filters['types'] = $this->getTypesProperty($propertyParameter);
+        $filters['excluded'] = $this->getExcludedFilter($filters, $propertyParameter);
 
         $queryResult = $this->getSearchResult($filters, $limit, $page, $pageSize, $options['locale']);
 
         $result = [];
-        $uuids = [];
         /** @var ArticleViewDocumentInterface $document */
         foreach ($queryResult as $document) {
-            $uuids[] = $document->getUuid();
             $result[] = new ArticleDataItem($document->getUuid(), $document->getTitle(), $document);
         }
 
-        return new DataProviderResult($result, $this->hasNextPage($queryResult, $limit, $page, $pageSize), $uuids);
+        return new DataProviderResult($result, $this->hasNextPage($queryResult, $limit, $page, $pageSize));
     }
 
     /**
@@ -141,21 +175,18 @@ class ArticleDataProvider implements DataProviderInterface
         $pageSize = null
     ) {
         $filters['types'] = $this->getTypesProperty($propertyParameter);
+        $filters['excluded'] = $this->getExcludedFilter($filters, $propertyParameter);
 
         $queryResult = $this->getSearchResult($filters, $limit, $page, $pageSize, $options['locale']);
 
         $result = [];
-        $uuids = [];
         /** @var ArticleViewDocumentInterface $document */
         foreach ($queryResult as $document) {
-            $uuids[] = $document->getUuid();
-            $result[] = new ArticleResourceItem(
-                $document,
-                $this->getResource($document->getUuid(), $document->getLocale())
-            );
+            $this->referenceStore->add($document->getUuid());
+            $result[] = $this->articleResourceItemFactory->createResourceItem($document);
         }
 
-        return new DataProviderResult($result, $this->hasNextPage($queryResult, $limit, $page, $pageSize), $uuids);
+        return new DataProviderResult($result, $this->hasNextPage($queryResult, $limit, $page, $pageSize));
     }
 
     /**
@@ -170,14 +201,14 @@ class ArticleDataProvider implements DataProviderInterface
      * Returns flag "hasNextPage".
      * It combines the limit/query-count with the page and page-size.
      *
-     * @param DocumentIterator $queryResult
+     * @param \Countable $queryResult
      * @param int $limit
      * @param int $page
      * @param int $pageSize
      *
      * @return bool
      */
-    private function hasNextPage(DocumentIterator $queryResult, $limit, $page, $pageSize)
+    private function hasNextPage(\Countable $queryResult, $limit, $page, $pageSize)
     {
         $count = $queryResult->count();
         if ($limit && $limit < $count) {
@@ -196,13 +227,42 @@ class ArticleDataProvider implements DataProviderInterface
      * @param int $pageSize
      * @param string $locale
      *
-     * @return DocumentIterator
+     * @return \Countable
      */
     private function getSearchResult(array $filters, $limit, $page, $pageSize, $locale)
     {
         $repository = $this->searchManager->getRepository($this->articleDocumentClass);
-        /** @var Search $search */
-        $search = $repository->createSearch();
+        $search = $this->createSearch($repository->createSearch(), $filters, $locale);
+        if (!$search) {
+            return new \ArrayIterator([]);
+        }
+
+        $this->addPagination($search, $pageSize, $page, $limit);
+
+        if (array_key_exists('sortBy', $filters) && is_array($filters['sortBy'])) {
+            $sortMethod = array_key_exists('sortMethod', $filters) ? $filters['sortMethod'] : 'asc';
+            $this->appendSortBy($filters['sortBy'], $sortMethod, $search);
+        }
+
+        return $repository->findDocuments($search);
+    }
+
+    /**
+     * Initialize search with neccesary queries.
+     *
+     * @param Search $search
+     * @param array $filters
+     * @param string $locale
+     *
+     * @return Search
+     */
+    protected function createSearch(Search $search, array $filters, $locale)
+    {
+        if (0 < count($filters['excluded'])) {
+            foreach ($filters['excluded'] as $uuid) {
+                $search->addQuery(new TermQuery('uuid', $uuid), BoolQuery::MUST_NOT);
+            }
+        }
 
         $query = new BoolQuery();
 
@@ -230,26 +290,12 @@ class ArticleDataProvider implements DataProviderInterface
         }
 
         if (0 === $queriesCount) {
-            $search->addQuery(new MatchAllQuery());
+            $search->addQuery(new MatchAllQuery(), BoolQuery::MUST);
         } else {
-            $search->addQuery($query);
+            $search->addQuery($query, BoolQuery::MUST);
         }
 
-        if (null !== $pageSize) {
-            $this->addPagination($search, $pageSize, $page, $limit);
-        } elseif (null !== $limit) {
-            $search->setSize($limit);
-        } else {
-            // FIXME find better way to achieve this
-            $search->setSize(1000);
-        }
-
-        if (array_key_exists('sortBy', $filters) && is_array($filters['sortBy'])) {
-            $sortMethod = array_key_exists('sortMethod', $filters) ? $filters['sortMethod'] : 'asc';
-            $this->appendSortBy($filters['sortBy'], $sortMethod, $search);
-        }
-
-        return $repository->findDocuments($search);
+        return $search;
     }
 
     /**
@@ -272,6 +318,26 @@ class ArticleDataProvider implements DataProviderInterface
         }
 
         return $filterTypes;
+    }
+
+    /**
+     * Returns excluded articles.
+     *
+     * @param array $filters
+     * @param PropertyParameter[] $propertyParameter
+     *
+     * @return array
+     */
+    private function getExcludedFilter(array $filters, array $propertyParameter)
+    {
+        $excluded = array_key_exists('excluded', $filters) ? $filters['excluded'] : [];
+        if (array_key_exists('exclude_duplicates', $propertyParameter)
+            && $propertyParameter['exclude_duplicates']->getValue()
+        ) {
+            $excluded = array_merge($excluded, $this->referenceStore->getAll());
+        }
+
+        return $excluded;
     }
 
     /**
@@ -300,19 +366,26 @@ class ArticleDataProvider implements DataProviderInterface
      */
     private function addPagination(Search $search, $pageSize, $page, $limit)
     {
-        $pageSize = intval($pageSize);
-        $offset = ($page - 1) * $pageSize;
-
-        $position = $pageSize * $page;
-        if ($limit !== null && $position >= $limit) {
-            $pageSize = $limit - $offset;
-            $loadLimit = $pageSize;
-        } else {
-            $loadLimit = $pageSize;
+        $offset = 0;
+        if ($pageSize) {
+            $pageSize = intval($pageSize);
+            $offset = ($page - 1) * $pageSize;
         }
 
-        $search->setSize($loadLimit);
+        if (null === $limit) {
+            $limit = $this->defaultLimit;
+        }
+
+        if (null === $pageSize || $offset + $pageSize > $limit) {
+            $pageSize = $limit - $offset;
+
+            if ($pageSize < 0) {
+                $pageSize = 0;
+            }
+        }
+
         $search->setFrom($offset);
+        $search->setSize($pageSize);
     }
 
     /**
@@ -410,5 +483,13 @@ class ArticleDataProvider implements DataProviderInterface
                 return true;
             }
         );
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getAlias()
+    {
+        return 'article';
     }
 }
