@@ -15,6 +15,7 @@ use ONGR\ElasticsearchBundle\Annotation\Document;
 use ONGR\ElasticsearchBundle\Annotation\Embedded;
 use ONGR\ElasticsearchBundle\Annotation\Id;
 use ONGR\ElasticsearchBundle\Annotation\Property;
+use ONGR\ElasticsearchBundle\Collection\Collection;
 
 /**
  * Indexable document for articles.
@@ -65,6 +66,7 @@ class ArticleViewDocument implements ArticleViewDocumentInterface
      * @Property(
      *     type="string",
      *     options={
+     *         "analyzer": "pathAnalyzer",
      *         "fields":{
      *            "raw":{"type":"string", "index":"not_analyzed"},
      *            "value":{"type":"string"}
@@ -73,6 +75,13 @@ class ArticleViewDocument implements ArticleViewDocumentInterface
      * )
      */
     protected $routePath;
+
+    /**
+     * @var string
+     *
+     * @Property(type="string", options={"analyzer": "keyword"})
+     */
+    protected $parentPageUuid;
 
     /**
      * @var string
@@ -247,11 +256,34 @@ class ArticleViewDocument implements ArticleViewDocumentInterface
     protected $changerContactId;
 
     /**
+     * @var ArticlePageViewObject[]
+     *
+     * @Embedded(class="SuluArticleBundle:ArticlePageViewObject", multiple=true)
+     */
+    protected $pages = [];
+
+    /**
+     * @var string
+     *
+     * @Property(type="binary")
+     */
+    protected $contentData;
+
+    /**
+     * @var \ArrayObject
+     */
+    protected $content;
+
+    /**
+     * @var \ArrayObject
+     */
+    protected $view;
+
+    /**
      * @param string $uuid
      */
-    public function __construct(
-        $uuid = null
-    ) {
+    public function __construct($uuid = null)
+    {
         $this->uuid = $uuid;
     }
 
@@ -341,6 +373,24 @@ class ArticleViewDocument implements ArticleViewDocumentInterface
     public function setRoutePath($routePath)
     {
         $this->routePath = $routePath;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getParentPageUuid()
+    {
+        return $this->parentPageUuid;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setParentPageUuid($parentPageUuid)
+    {
+        $this->parentPageUuid = $parentPageUuid;
+
+        return $this;
     }
 
     /**
@@ -681,5 +731,77 @@ class ArticleViewDocument implements ArticleViewDocumentInterface
     public function getChangerContactId()
     {
         return $this->changerContactId;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getPages()
+    {
+        return $this->pages;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setPages(Collection $pages)
+    {
+        $this->pages = $pages;
+
+        return $this;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getContentData()
+    {
+        return $this->contentData;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setContentData($contentData)
+    {
+        $this->contentData = $contentData;
+
+        return $this;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getContent()
+    {
+        return $this->content;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setContent(\ArrayObject $content)
+    {
+        $this->content = $content;
+
+        return $this;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getView()
+    {
+        return $this->view;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setView(\ArrayObject $view)
+    {
+        $this->view = $view;
+
+        return $this;
     }
 }
