@@ -12,18 +12,28 @@
 namespace Sulu\Bundle\ArticleBundle\Document\Subscriber;
 
 use Doctrine\ORM\EntityManagerInterface;
-use Sulu\Bundle\ArticleBundle\Document\ArticleDocument;
+use PHPCR\ItemNotFoundException;
+use PHPCR\SessionInterface;
 use Sulu\Bundle\ArticleBundle\Document\Behavior\RoutableBehavior;
+use Sulu\Bundle\ArticleBundle\Document\Behavior\RoutablePageBehavior;
 use Sulu\Bundle\DocumentManagerBundle\Bridge\DocumentInspector;
 use Sulu\Bundle\DocumentManagerBundle\Bridge\PropertyEncoder;
 use Sulu\Bundle\RouteBundle\Entity\RouteRepositoryInterface;
+use Sulu\Bundle\RouteBundle\Exception\RouteIsNotUniqueException;
+use Sulu\Bundle\RouteBundle\Generator\ChainRouteGeneratorInterface;
+use Sulu\Bundle\RouteBundle\Manager\ConflictResolverInterface;
 use Sulu\Bundle\RouteBundle\Manager\RouteManagerInterface;
+use Sulu\Bundle\RouteBundle\Model\RouteInterface;
+use Sulu\Component\Content\Exception\ResourceLocatorAlreadyExistsException;
+use Sulu\Component\Content\Metadata\Factory\StructureMetadataFactoryInterface;
+use Sulu\Component\DocumentManager\Behavior\Mapping\ChildrenBehavior;
+use Sulu\Component\DocumentManager\Behavior\Mapping\ParentBehavior;
 use Sulu\Component\DocumentManager\DocumentManagerInterface;
 use Sulu\Component\DocumentManager\Event\AbstractMappingEvent;
-use Sulu\Component\DocumentManager\Event\ConfigureOptionsEvent;
 use Sulu\Component\DocumentManager\Event\CopyEvent;
-use Sulu\Component\DocumentManager\Event\MetadataLoadEvent;
+use Sulu\Component\DocumentManager\Event\PublishEvent;
 use Sulu\Component\DocumentManager\Event\RemoveEvent;
+use Sulu\Component\DocumentManager\Event\ReorderEvent;
 use Sulu\Component\DocumentManager\Events;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
@@ -33,6 +43,15 @@ use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 class RoutableSubscriber implements EventSubscriberInterface
 {
     const ROUTE_FIELD = 'routePath';
+
+    const ROUTES_PROPERTY = 'suluRoutes';
+
+    const TAG_NAME = 'sulu_article.article_route';
+
+    /**
+     * @var ChainRouteGeneratorInterface
+     */
+    private $chainRouteGenerator;
 
     /**
      * @var RouteManagerInterface
@@ -65,27 +84,46 @@ class RoutableSubscriber implements EventSubscriberInterface
     private $propertyEncoder;
 
     /**
+     * @var StructureMetadataFactoryInterface
+     */
+    private $metadataFactory;
+
+    /**
+     * @var ConflictResolverInterface
+     */
+    private $conflictResolver;
+
+    /**
+     * @param ChainRouteGeneratorInterface $chainRouteGenerator
      * @param RouteManagerInterface $routeManager
      * @param RouteRepositoryInterface $routeRepository
      * @param EntityManagerInterface $entityManager
      * @param DocumentManagerInterface $documentManager
      * @param DocumentInspector $documentInspector
      * @param PropertyEncoder $propertyEncoder
+     * @param StructureMetadataFactoryInterface $metadataFactory
+     * @param ConflictResolverInterface $conflictResolver
      */
     public function __construct(
+        ChainRouteGeneratorInterface $chainRouteGenerator,
         RouteManagerInterface $routeManager,
         RouteRepositoryInterface $routeRepository,
         EntityManagerInterface $entityManager,
         DocumentManagerInterface $documentManager,
         DocumentInspector $documentInspector,
-        PropertyEncoder $propertyEncoder
+        PropertyEncoder $propertyEncoder,
+        StructureMetadataFactoryInterface $metadataFactory,
+        ConflictResolverInterface $conflictResolver
     ) {
+        $this->chainRouteGenerator = $chainRouteGenerator;
         $this->routeManager = $routeManager;
         $this->routeRepository = $routeRepository;
         $this->entityManager = $entityManager;
         $this->documentManager = $documentManager;
         $this->documentInspector = $documentInspector;
         $this->propertyEncoder = $propertyEncoder;
+        $this->metadataFactory = $metadataFactory;
+        $this->conflictResolver = $conflictResolver;
     }
 
     /**
@@ -94,12 +132,18 @@ class RoutableSubscriber implements EventSubscriberInterface
     public static function getSubscribedEvents()
     {
         return [
-            Events::HYDRATE => [['handleHydrate', -500]],
-            Events::PERSIST => [['handleRouteUpdate', 1], ['handleRoute', 0]],
-            Events::REMOVE => [['handleRemove', -500]],
+            Events::HYDRATE => ['handleHydrate'],
+            Events::PERSIST => [
+                // low priority because all other subscriber should be finished
+                ['handlePersist', -2000],
+            ],
+            Events::REMOVE => [
+                // high priority to ensure nodes are not deleted until we iterate over children
+                ['handleRemove', 1024],
+            ],
+            Events::PUBLISH => ['handlePublish', -2000],
+            Events::REORDER => ['handleReorder', -1000],
             Events::COPY => ['handleCopy', -2000],
-            Events::METADATA_LOAD => 'handleMetadataLoad',
-            Events::CONFIGURE_OPTIONS => 'configureOptions',
         ];
     }
 
@@ -111,56 +155,286 @@ class RoutableSubscriber implements EventSubscriberInterface
     public function handleHydrate(AbstractMappingEvent $event)
     {
         $document = $event->getDocument();
-        if (!$document instanceof RoutableBehavior || null === $document->getRoutePath()) {
+        if (!$document instanceof RoutablePageBehavior) {
             return;
         }
 
-        $route = $this->routeRepository->findByPath($document->getRoutePath(), $document->getOriginalLocale());
-        if (!$route) {
-            return;
-        }
+        $propertyName = $this->getRoutePathPropertyName($document->getStructureType(), $event->getLocale());
+        $routePath = $event->getNode()->getPropertyValueWithDefault($propertyName, null);
+        $document->setRoutePath($routePath);
 
-        $document->setRoute($route);
+        $route = $this->routeRepository->findByEntity($document->getClass(), $document->getUuid(), $event->getLocale());
+        if ($route) {
+            $document->setRoute($route);
+        }
     }
 
     /**
-     * Generate route.
+     * Generate route and save route-path.
      *
      * @param AbstractMappingEvent $event
      */
-    public function handleRoute(AbstractMappingEvent $event)
+    public function handlePersist(AbstractMappingEvent $event)
     {
         $document = $event->getDocument();
-        if (!$document instanceof RoutableBehavior || null !== $document->getRoutePath()) {
+        if (!$document instanceof RoutablePageBehavior) {
             return;
         }
 
         $document->setUuid($event->getNode()->getIdentifier());
 
-        $route = $this->routeManager->create($document, $event->getOption('route_path'));
+        $propertyName = $this->getRoutePathPropertyName($document->getStructureType(), $event->getLocale());
+        $routePath = $event->getNode()->getPropertyValueWithDefault($propertyName, null);
+
+        $route = $this->conflictResolver->resolve($this->chainRouteGenerator->generate($document, $routePath));
+        $document->setRoutePath($route->getPath());
+
+        $event->getNode()->setProperty($propertyName, $route->getPath());
+
+        if (!$document instanceof ChildrenBehavior) {
+            return;
+        }
+
+        foreach ($document->getChildren() as $child) {
+            if (!$child instanceof RoutablePageBehavior) {
+                continue;
+            }
+
+            $route = $this->chainRouteGenerator->generate($child);
+            $child->setRoutePath($route->getPath());
+
+            $node = $this->documentInspector->getNode($child);
+            $node->setProperty($propertyName, $route->getPath());
+        }
+    }
+
+    /**
+     * Regenerate routes for siblings on reorder.
+     *
+     * @param ReorderEvent $event
+     */
+    public function handleReorder(ReorderEvent $event)
+    {
+        $document = $event->getDocument();
+        if (!$document instanceof RoutablePageBehavior || !$document instanceof ParentBehavior) {
+            return;
+        }
+
+        $parentDocument = $document->getParent();
+        if (!$parentDocument instanceof ChildrenBehavior) {
+            return;
+        }
+
+        $locale = $this->documentInspector->getLocale($parentDocument);
+        $propertyName = $this->getRoutePathPropertyName($parentDocument->getStructureType(), $locale);
+        foreach ($parentDocument->getChildren() as $childDocument) {
+            $node = $this->documentInspector->getNode($childDocument);
+
+            $route = $this->chainRouteGenerator->generate($childDocument);
+            $childDocument->setRoutePath($route->getPath());
+
+            $node->setProperty($propertyName, $route->getPath());
+        }
+    }
+
+    /**
+     * Handle publish event and generate route and the child-routes.
+     *
+     * @param PublishEvent $event
+     *
+     * @throws ResourceLocatorAlreadyExistsException
+     */
+    public function handlePublish(PublishEvent $event)
+    {
+        $document = $event->getDocument();
+        if (!$document instanceof RoutableBehavior) {
+            return;
+        }
+
+        $node = $this->documentInspector->getNode($document);
+
+        try {
+            $route = $this->createOrUpdateRoute($document, $event->getLocale());
+        } catch (RouteIsNotUniqueException $exception) {
+            throw new ResourceLocatorAlreadyExistsException($exception->getRoute()->getPath(), $document->getPath());
+        }
+
+        $document->setRoutePath($route->getPath());
         $this->entityManager->persist($route);
+
+        $node->setProperty(
+            $this->getRoutePathPropertyName($document->getStructureType(), $event->getLocale()),
+            $route->getPath()
+        );
+
+        $propertyName = $this->getPropertyName($event->getLocale(), self::ROUTES_PROPERTY);
+
+        // check if nodes previous generated routes exists and remove them if not
+        $oldRoutes = $event->getNode()->getPropertyValueWithDefault($propertyName, []);
+        $this->removeOldChildRoutes($event->getNode()->getSession(), $oldRoutes, $event->getLocale());
+
+        $routes = [];
+        if ($document instanceof ChildrenBehavior) {
+            // generate new routes of children
+            $routes = $this->generateChildRoutes($document, $event->getLocale());
+        }
+
+        // save the newly generated routes of children
+        $event->getNode()->setProperty($propertyName, $routes);
         $this->entityManager->flush();
     }
 
     /**
-     * Update route if route-path was changed.
+     * Create or update for given document.
      *
-     * @param AbstractMappingEvent $event
+     * @param RoutablePageBehavior $document
+     * @param string $locale
+     *
+     * @return RouteInterface
      */
-    public function handleRouteUpdate(AbstractMappingEvent $event)
+    private function createOrUpdatePageRoute(RoutablePageBehavior $document, $locale)
     {
-        $document = $event->getDocument();
-        if (!$document instanceof RoutableBehavior
-            || null === $document->getRoute()
-            || null === ($routePath = $event->getOption('route_path'))
-        ) {
+        $route = $this->reallocateExistingRoute($document, $locale);
+        if ($route) {
+            return $route;
+        }
+
+        $route = $document->getRoute();
+        if (!$route) {
+            $route = $this->routeRepository->findByEntity($document->getClass(), $document->getUuid(), $locale);
+        }
+
+        if ($route && $route->getEntityId() !== $document->getId()) {
+            // Mismatch of entity-id's happens because doctrine don't check entities which has been changed in the
+            // current session.
+
+            $document->removeRoute();
+            $route = null;
+        }
+
+        if ($route) {
+            $document->setRoute($route);
+
+            return $this->routeManager->update($document, null, false);
+        }
+
+        return $this->routeManager->create($document);
+    }
+
+    /**
+     * Reallocates existing route to given document.
+     *
+     * @param RoutablePageBehavior $document
+     * @param string $locale
+     *
+     * @return RouteInterface
+     */
+    private function reallocateExistingRoute(RoutablePageBehavior $document, $locale)
+    {
+        $newRoute = $this->routeRepository->findByPath($document->getRoutePath(), $locale);
+        if (!$newRoute) {
             return;
         }
 
-        $route = $this->routeManager->update($document, $routePath);
-        $document->setRoutePath($route->getPath());
-        $this->entityManager->persist($route);
+        $oldRoute = $this->routeRepository->findByEntity(get_class($document), $document->getUuid(), $locale);
+        $history = $this->routeRepository->findHistoryByEntity(get_class($document), $document->getUuid(), $locale);
+
+        /** @var RouteInterface $historyRoute */
+        foreach (array_filter(array_merge($history, [$oldRoute])) as $historyRoute) {
+            if ($historyRoute->getId() === $newRoute->getId() || $document->getId() !== $historyRoute->getEntityId()) {
+                // Mismatch of entity-id's happens because doctrine don't check entities which has been changed in the
+                // current session. If the old-route was already reused by a page before it will be returned in the
+                // query of line 329.
+
+                continue;
+            }
+
+            $historyRoute->setTarget($newRoute);
+            $historyRoute->setHistory(true);
+            $newRoute->addHistory($historyRoute);
+        }
+
+        $newRoute->setEntityClass(get_class($document));
+        $newRoute->setEntityId($document->getId());
+        $newRoute->setTarget(null);
+        $newRoute->setHistory(false);
+
+        return $newRoute;
+    }
+
+    /**
+     * Create or update for given document.
+     *
+     * @param RoutableBehavior $document
+     * @param string $locale
+     *
+     * @return RouteInterface
+     */
+    private function createOrUpdateRoute(RoutableBehavior $document, $locale)
+    {
+        $route = $document->getRoute();
+
+        if (!$route) {
+            $route = $this->routeRepository->findByEntity($document->getClass(), $document->getUuid(), $locale);
+        }
+
+        if ($route) {
+            $document->setRoute($route);
+
+            return $this->routeManager->update($document, $document->getRoutePath(), false);
+        }
+
+        return $this->routeManager->create($document, $document->getRoutePath(), false);
+    }
+
+    /**
+     * Removes old-routes where the node does not exists anymore.
+     *
+     * @param SessionInterface $session
+     * @param array $oldRoutes
+     * @param string $locale
+     */
+    private function removeOldChildRoutes(SessionInterface $session, array $oldRoutes, $locale)
+    {
+        foreach ($oldRoutes as $oldRoute) {
+            $oldRouteEntity = $this->routeRepository->findByPath($oldRoute, $locale);
+            if ($oldRouteEntity && !$this->nodeExists($session, $oldRouteEntity->getEntityId())) {
+                $this->entityManager->remove($oldRouteEntity);
+            }
+        }
+
         $this->entityManager->flush();
+    }
+
+    /**
+     * Generates child routes.
+     *
+     * @param ChildrenBehavior $document
+     * @param string $locale
+     *
+     * @return string[]
+     */
+    private function generateChildRoutes(ChildrenBehavior $document, $locale)
+    {
+        $routes = [];
+        foreach ($document->getChildren() as $child) {
+            if (!$child instanceof RoutablePageBehavior) {
+                continue;
+            }
+
+            $childRoute = $this->createOrUpdatePageRoute($child, $locale);
+            $this->entityManager->persist($childRoute);
+
+            $child->setRoutePath($childRoute->getPath());
+            $childNode = $this->documentInspector->getNode($child);
+
+            $propertyName = $this->getRoutePathPropertyName($child->getStructureType(), $locale);
+            $childNode->setProperty($propertyName, $childRoute->getPath());
+
+            $routes[] = $childRoute->getPath();
+        }
+
+        return $routes;
     }
 
     /**
@@ -175,12 +449,26 @@ class RoutableSubscriber implements EventSubscriberInterface
             return;
         }
 
-        $route = $this->routeRepository->findByPath($document->getRoutePath(), $document->getOriginalLocale());
-        if (!$route) {
-            return;
+        $locales = $this->documentInspector->getLocales($document);
+        foreach ($locales as $locale) {
+            $localizedDocument = $this->documentManager->find($document->getUuid(), $locale);
+
+            $route = $this->routeRepository->findByEntity(
+                $localizedDocument->getClass(),
+                $localizedDocument->getUuid(),
+                $locale
+            );
+            if (!$route) {
+                continue;
+            }
+
+            $this->entityManager->remove($route);
+
+            if ($document instanceof ChildrenBehavior) {
+                $this->removeChildRoutes($document, $locale);
+            }
         }
 
-        $this->entityManager->remove($route);
         $this->entityManager->flush();
     }
 
@@ -198,52 +486,108 @@ class RoutableSubscriber implements EventSubscriberInterface
 
         $locales = $this->documentInspector->getLocales($document);
         foreach ($locales as $locale) {
-            /** @var ArticleDocument $localizedDocument */
             $localizedDocument = $this->documentManager->find($event->getCopiedPath(), $locale);
 
-            $localizedDocument->removeRoute();
-            $route = $this->routeManager->create($localizedDocument);
-            $document->setRoutePath($route->getPath());
-            $this->entityManager->persist($route);
+            $route = $this->conflictResolver->resolve($this->chainRouteGenerator->generate($localizedDocument));
+            $localizedDocument->setRoutePath($route->getPath());
 
-            $event->getCopiedNode()->setProperty(
-                $this->propertyEncoder->localizedSystemName(self::ROUTE_FIELD, $locale),
+            $node = $this->documentInspector->getNode($localizedDocument);
+            $node->setProperty(
+                $this->getRoutePathPropertyName($localizedDocument->getStructureType(), $locale),
                 $route->getPath()
             );
-        }
 
-        $this->entityManager->flush();
+            $propertyName = $this->getRoutePathPropertyName($localizedDocument->getStructureType(), $locale);
+            $node = $this->documentInspector->getNode($localizedDocument);
+            $node->setProperty($propertyName, $route->getPath());
+
+            if ($localizedDocument instanceof ChildrenBehavior) {
+                $this->generateChildRoutes($localizedDocument, $locale);
+            }
+        }
     }
 
     /**
-     * Add route to metadata.
+     * Iterate over children and remove routes.
      *
-     * @param MetadataLoadEvent $event
+     * @param ChildrenBehavior $document
+     * @param string $locale
      */
-    public function handleMetadataLoad(MetadataLoadEvent $event)
+    private function removeChildRoutes(ChildrenBehavior $document, $locale)
     {
-        if (!$event->getMetadata()->getReflectionClass()->implementsInterface(RoutableBehavior::class)) {
-            return;
-        }
+        foreach ($document->getChildren() as $child) {
+            if ($child instanceof RoutablePageBehavior) {
+                $this->removeChildRoute($child, $locale);
+            }
 
-        $metadata = $event->getMetadata();
-        $metadata->addFieldMapping(
-            self::ROUTE_FIELD,
-            [
-                'encoding' => 'system_localized',
-                'property' => self::ROUTE_FIELD,
-            ]
-        );
+            if ($child instanceof ChildrenBehavior) {
+                $this->removeChildRoutes($child, $locale);
+            }
+        }
     }
 
     /**
-     * Add route-path to options.
+     * Removes route if exists.
      *
-     * @param ConfigureOptionsEvent $event
+     * @param RoutablePageBehavior $document
+     * @param string $locale
      */
-    public function configureOptions(ConfigureOptionsEvent $event)
+    private function removeChildRoute(RoutablePageBehavior $document, $locale)
     {
-        $options = $event->getOptions();
-        $options->setDefaults(['route_path' => null]);
+        $route = $this->routeRepository->findByPath($document->getRoutePath(), $locale);
+        if ($route) {
+            $this->entityManager->remove($route);
+        }
+    }
+
+    /**
+     * Returns encoded "routePath" property-name.
+     *
+     * @param string $structureType
+     * @param string $locale
+     *
+     * @return string
+     */
+    private function getRoutePathPropertyName($structureType, $locale)
+    {
+        $metadata = $this->metadataFactory->getStructureMetadata('article', $structureType);
+
+        if ($metadata->hasTag(self::TAG_NAME)) {
+            return $this->getPropertyName($locale, $metadata->getPropertyByTagName(self::TAG_NAME)->getName());
+        }
+
+        return $this->getPropertyName($locale, self::ROUTE_FIELD);
+    }
+
+    /**
+     * Returns encoded property-name.
+     *
+     * @param string $locale
+     * @param string $field
+     *
+     * @return string
+     */
+    private function getPropertyName($locale, $field)
+    {
+        return $this->propertyEncoder->localizedSystemName($field, $locale);
+    }
+
+    /**
+     * Returns true if given uuid exists.
+     *
+     * @param SessionInterface $session
+     * @param string $uuid
+     *
+     * @return bool
+     */
+    private function nodeExists(SessionInterface $session, $uuid)
+    {
+        try {
+            $session->getNodeByIdentifier($uuid);
+
+            return true;
+        } catch (ItemNotFoundException $exception) {
+            return false;
+        }
     }
 }
