@@ -15,6 +15,7 @@ use JMS\Serializer\DeserializationContext;
 use JMS\Serializer\SerializationContext;
 use JMS\Serializer\SerializerInterface;
 use Sulu\Bundle\ArticleBundle\Document\ArticleDocument;
+use Sulu\Bundle\ArticleBundle\Document\ArticlePageDocument;
 use Sulu\Bundle\PreviewBundle\Preview\Object\PreviewObjectProviderInterface;
 use Sulu\Component\DocumentManager\DocumentManagerInterface;
 use Symfony\Component\PropertyAccess\PropertyAccess;
@@ -35,13 +36,23 @@ class ArticleObjectProvider implements PreviewObjectProviderInterface
     private $serializer;
 
     /**
+     * @var string
+     */
+    private $articleDocumentClass;
+
+    /**
      * @param DocumentManagerInterface $documentManager
      * @param SerializerInterface $serializer
+     * @param $articleDocumentClass
      */
-    public function __construct(DocumentManagerInterface $documentManager, SerializerInterface $serializer)
-    {
+    public function __construct(
+        DocumentManagerInterface $documentManager,
+        SerializerInterface $serializer,
+        $articleDocumentClass
+    ) {
         $this->documentManager = $documentManager;
         $this->serializer = $serializer;
+        $this->articleDocumentClass = $articleDocumentClass;
     }
 
     /**
@@ -49,7 +60,14 @@ class ArticleObjectProvider implements PreviewObjectProviderInterface
      */
     public function getObject($id, $locale)
     {
-        return $this->documentManager->find($id, $locale);
+        return $this->documentManager->find(
+            $id,
+            $locale,
+            [
+                'load_ghost_content' => false,
+                'load_shadow_content' => true,
+            ]
+        );
     }
 
     /**
@@ -104,11 +122,23 @@ class ArticleObjectProvider implements PreviewObjectProviderInterface
      */
     public function serialize($object)
     {
-        return $this->serializer->serialize(
+        $pageNumber = $object->getPageNumber();
+        if ($object instanceof ArticlePageDocument) {
+            // resolve proxy to ensure that this will also be serialized
+            $object = $object->getParent();
+
+            $object->getTitle();
+        }
+
+        $result = $this->serializer->serialize(
             $object,
             'json',
-            SerializationContext::create()->setSerializeNull(true)->setGroups(['preview'])
+            SerializationContext::create()
+                ->setSerializeNull(true)
+                ->setGroups(['preview'])
         );
+
+        return json_encode(['pageNumber' => $pageNumber, 'object' => $result]);
     }
 
     /**
@@ -118,11 +148,29 @@ class ArticleObjectProvider implements PreviewObjectProviderInterface
      */
     public function deserialize($serializedObject, $objectClass)
     {
-        return $this->serializer->deserialize(
-            $serializedObject,
-            $objectClass,
+        $result = json_decode($serializedObject, true);
+
+        $article = $this->serializer->deserialize(
+            $result['object'],
+            $this->articleDocumentClass,
             'json',
-            DeserializationContext::create()->setSerializeNull(true)->setGroups(['preview'])
+            DeserializationContext::create()
+                ->setSerializeNull(true)
+                ->setGroups(['preview'])
         );
+
+        foreach ($article->getChildren() as $child) {
+            $child->setParent($article);
+        }
+
+        if (1 === $result['pageNumber']) {
+            return $article;
+        }
+
+        $children = array_values($article->getChildren());
+
+        $object = $children[$result['pageNumber'] - 2];
+
+        return $object;
     }
 }
