@@ -10,8 +10,10 @@
 define([
     'underscore',
     'jquery',
-    'services/suluarticle/article-manager'
-], function(_, $, ArticleManager) {
+    'services/suluarticle/article-manager',
+    'services/suluarticle/article-router',
+    'services/suluarticle/property-configuration'
+], function(_, $, ArticleManager, ArticleRouter, PropertyConfiguration) {
 
     'use strict';
 
@@ -53,6 +55,7 @@ define([
 
         listenForChange: function() {
             this.sandbox.dom.on(this.$el, 'keyup', _.debounce(this.setDirty.bind(this), 10), 'input, textarea');
+            this.sandbox.dom.on(this.$el, 'change', _.debounce(this.setDirty.bind(this), 10), 'input, textarea');
             this.sandbox.dom.on(this.$el, 'change', _.debounce(this.setDirty.bind(this), 10), 'input[type="checkbox"], select');
             this.sandbox.on('sulu.content.changed', this.setDirty.bind(this));
         },
@@ -72,30 +75,29 @@ define([
                 return;
             }
 
+            var routePathProperty = this.getRoutePathProperty();
             var data = this.sandbox.form.getData(this.formId);
-            data.template = this.template;
 
-            _.each(data, function(value, key) {
-                this.data[key] = value;
-            }.bind(this));
+            if (data[routePathProperty.name] && data[routePathProperty.name] === '/') {
+                delete data[routePathProperty.name];
+            }
 
-            ArticleManager.save(this.data, this.data.id, this.options.locale, action).then(function(response) {
+            this.options.adapter.save(this, data, action).done(function(response) {
                 this.data = response;
 
                 if (this.ghost && !this.data.type) {
-                    this.sandbox.emit(
+                    // reload page
+                    return this.sandbox.emit(
                         'sulu.router.navigate',
-                        'articles/' + this.options.locale + '/edit:' + this.data.id + '/details',
+                        this.sandbox.mvc.history.fragment,
                         true,
                         true
                     );
-
-                    return;
                 }
 
                 this.sandbox.emit('sulu.tab.saved', response.id, response);
             }.bind(this)).fail(function(xhr) {
-                this.sandbox.emit('sulu.article.error', xhr.status, data);
+                this.sandbox.emit('sulu.article.error', xhr.status, xhr.responseJSON.code || 0, data);
             }.bind(this));
         },
 
@@ -166,9 +168,9 @@ define([
                 this.data = this.sandbox.util.extend({}, data, this.data);
             }
 
-            require([this.getTemplateUrl(template)], function(template) {
-                this.renderFormTemplate(template);
-            }.bind(this));
+            this.sandbox.emit('sulu.article.update-page-switcher', template);
+
+            require([this.getTemplateUrl(template)], this.renderFormTemplate.bind(this));
         },
 
         /**
@@ -199,30 +201,32 @@ define([
             this.sandbox.dom.html(this.formId, this.sandbox.util.template(template, {
                 translate: this.sandbox.translate,
                 content: this.data,
-                options: this.options
+                options: this.options,
+                categoryLocale: this.options.locale,
+                entityClass: 'Sulu\\Bundle\\ArticleBundle\\Document\\ArticleDocument',
+                entityId: this.options.id
             }));
 
-            if (!this.data.id || (this.data.type && this.data.type.name === 'ghost')) {
-                // route-path will be generator on post-request
-                this.$find('#routePath').parent().remove();
-                this.data.routePath = null;
-            }
+            var data = this.options.adapter.prepareData(this.data, this);
+            if (data.type && data.type.name === 'ghost') {
+                var titleProperty = this.getTitleProperty(),
+                    pageTitleProperty = this.getPageTitleProperty();
 
-            if (this.data.type && this.data.type.name === 'ghost') {
                 this.ghost = {
-                    locale: this.data.type.value,
-                    title: this.data.title
+                    locale: data.type.value,
+                    title: titleProperty ? data[titleProperty.name] : '',
+                    pageTitle: pageTitleProperty ? data[pageTitleProperty.name] : ''
                 };
 
-                this.data = {
+                data = {
                     id: this.data.id,
                     articleType: this.data.articleType
                 };
             }
 
-            this.createForm(this.data).then(function() {
-                this.changeTemplateDropdownHandler();
-            }.bind(this));
+            this.options.adapter.beforeFormCreate(this);
+
+            this.createForm(data).then(this.changeTemplateDropdownHandler.bind(this));
         },
 
         changeTemplateDropdownHandler: function() {
@@ -244,7 +248,16 @@ define([
             formObject.initialized.then(function() {
                 this.sandbox.form.setData(this.formId, data).then(function() {
                     if (!!this.ghost) {
-                        this.sandbox.dom.attr('#title', 'placeholder', this.ghost.locale + ': ' + this.ghost.title);
+                        var titleProperty = this.getTitleProperty(),
+                            pageTitleProperty = this.getPageTitleProperty();
+
+                        if (titleProperty) {
+                            this.sandbox.dom.attr(titleProperty.$el, 'placeholder', this.ghost.locale + ': ' + this.ghost.title);
+                        }
+
+                        if (pageTitleProperty) {
+                            this.sandbox.dom.attr(pageTitleProperty.$el, 'placeholder', this.ghost.locale + ': ' + this.ghost.pageTitle);
+                        }
                     }
 
                     this.sandbox.start(this.$el, {reset: true}).then(function() {
@@ -252,11 +265,20 @@ define([
                         this.bindFormEvents();
                         deferred.resolve();
 
+                        this.sandbox.emit('sulu.content.initialized');
+
                         if (!!this.options.preview) {
                             this.options.preview.bindDomEvents(this.$el);
-                            var data = this.data;
+                            var data = this.options.adapter.prepareData(this.data, this);
                             data.template = this.template;
-                            this.options.preview.updateContext({template: this.template}, data);
+                            if (!!data.type && data.type.name === 'ghost') {
+                                data = {id: data.id};
+                            }
+
+                            this.options.preview.updateContext(
+                                {template: this.template},
+                                this.options.adapter.prepareData(data, this)
+                            );
                         }
                     }.bind(this));
                 }.bind(this));
@@ -321,11 +343,7 @@ define([
         },
 
         loadComponentData: function() {
-            var promise = $.Deferred();
-
-            promise.resolve(this.options.data());
-
-            return promise;
+            return this.options.data();
         },
 
         /**
@@ -341,6 +359,42 @@ define([
                 }
 
                 this.options.preview.updateProperty(propertyName, data[propertyName]);
+            }
+        },
+
+        getTitleProperty: function() {
+            if (!this.propertyConfiguration) {
+                this.propertyConfiguration = PropertyConfiguration.generate(this.$el);
+            }
+
+            if (!!this.propertyConfiguration.tags['sulu_article.article_title']) {
+                return this.propertyConfiguration.tags['sulu_article.article_title'].highestProperty;
+            }
+
+            return this.propertyConfiguration.title;
+        },
+
+        getPageTitleProperty: function() {
+            if (!this.propertyConfiguration) {
+                this.propertyConfiguration = PropertyConfiguration.generate(this.$el);
+            }
+
+            if (!!this.propertyConfiguration.tags['sulu_article.page_title']) {
+                return this.propertyConfiguration.tags['sulu_article.page_title'].highestProperty;
+            } else if (!!this.propertyConfiguration.pageTitle) {
+                return this.propertyConfiguration.pageTitle;
+            }
+        },
+
+        getRoutePathProperty: function() {
+            if (!this.propertyConfiguration) {
+                this.propertyConfiguration = PropertyConfiguration.generate(this.$el);
+            }
+
+            if (!!this.propertyConfiguration.tags['sulu_article.article_route']) {
+                return this.propertyConfiguration.tags['sulu_article.article_route'].highestProperty;
+            } else if (!!this.propertyConfiguration.routePath) {
+                return this.propertyConfiguration.routePath;
             }
         }
     };
