@@ -32,7 +32,7 @@ define([
             ArticleManager.save(content, content.id, this.options.locale, action).then(function(response) {
                 this.sandbox.emit('sulu.tab.saved', response.id, response);
             }.bind(this)).fail(function(xhr) {
-                this.sandbox.emit('sulu.article.error', xhr.status, data);
+                this.sandbox.emit('sulu.article.error', xhr.status, xhr.responseJSON.code || 0, data);
             }.bind(this));
         }
     };
