@@ -12,6 +12,8 @@
 namespace Sulu\Bundle\ArticleBundle\Routing;
 
 use Sulu\Bundle\ArticleBundle\Document\ArticleDocument;
+use Sulu\Bundle\ArticleBundle\Document\ArticleInterface;
+use Sulu\Bundle\ArticleBundle\Document\ArticlePageDocument;
 use Sulu\Bundle\RouteBundle\Routing\Defaults\RouteDefaultsProviderInterface;
 use Sulu\Component\Content\Compat\StructureManagerInterface;
 use Sulu\Component\Content\Document\WorkflowStage;
@@ -74,6 +76,12 @@ class ArticleRouteDefaultProvider implements RouteDefaultsProviderInterface
             $object = $this->documentManager->find($id, $locale);
         }
 
+        $pageNumber = $object->getPageNumber();
+        if ($object instanceof ArticlePageDocument) {
+            // the article contains the seo/excerpt data and the controller handles the page-number automatically
+            $object = $object->getParent();
+        }
+
         $metadata = $this->structureMetadataFactory->getStructureMetadata('article', $object->getStructureType());
 
         // this parameter should not be used
@@ -84,6 +92,7 @@ class ArticleRouteDefaultProvider implements RouteDefaultsProviderInterface
         return [
             'object' => $object,
             'view' => $metadata->view,
+            'pageNumber' => $pageNumber,
             'structure' => $structure,
             '_cacheLifetime' => $this->getCacheLifetime($metadata),
             '_controller' => $metadata->controller,
@@ -100,7 +109,7 @@ class ArticleRouteDefaultProvider implements RouteDefaultsProviderInterface
     {
         $object = $this->documentManager->find($id, $locale);
 
-        return $object instanceof ArticleDocument && WorkflowStage::PUBLISHED === $object->getWorkflowStage();
+        return $object instanceof ArticleInterface && WorkflowStage::PUBLISHED === $object->getWorkflowStage();
     }
 
     /**
@@ -108,7 +117,10 @@ class ArticleRouteDefaultProvider implements RouteDefaultsProviderInterface
      */
     public function supports($entityClass)
     {
-        return $entityClass === ArticleDocument::class;
+        return ArticleDocument::class === $entityClass
+            || ArticlePageDocument::class === $entityClass
+            || is_subclass_of($entityClass, ArticleDocument::class)
+            || is_subclass_of($entityClass, ArticlePageDocument::class);
     }
 
     /**
