@@ -13,10 +13,12 @@ namespace Sulu\Bundle\ArticleBundle\Teaser;
 
 use ONGR\ElasticsearchBundle\Service\Manager;
 use ONGR\ElasticsearchDSL\Query\TermLevel\IdsQuery;
+use Sulu\Bundle\ArticleBundle\Document\ArticleViewDocument;
 use Sulu\Bundle\ArticleBundle\Metadata\ArticleViewDocumentIdTrait;
 use Sulu\Bundle\ContentBundle\Teaser\Configuration\TeaserConfiguration;
 use Sulu\Bundle\ContentBundle\Teaser\Provider\TeaserProviderInterface;
 use Sulu\Bundle\ContentBundle\Teaser\Teaser;
+use Symfony\Component\Translation\TranslatorInterface;
 
 /**
  * Enables selection of articles in teaser content-type.
@@ -31,17 +33,24 @@ class ArticleTeaserProvider implements TeaserProviderInterface
     private $searchManager;
 
     /**
+     * @var TranslatorInterface
+     */
+    private $translator;
+
+    /**
      * @var string
      */
     private $articleDocumentClass;
 
     /**
      * @param Manager $searchManager
+     * @param TranslatorInterface $translator
      * @param $articleDocumentClass
      */
-    public function __construct(Manager $searchManager, $articleDocumentClass)
+    public function __construct(Manager $searchManager, TranslatorInterface $translator, $articleDocumentClass)
     {
         $this->searchManager = $searchManager;
+        $this->translator = $translator;
         $this->articleDocumentClass = $articleDocumentClass;
     }
 
@@ -50,22 +59,95 @@ class ArticleTeaserProvider implements TeaserProviderInterface
      */
     public function getConfiguration()
     {
+        $okDefaultText = $this->translator->trans('sulu-content.teaser.apply', [], 'backend');
+
         return new TeaserConfiguration(
             'sulu_article.teaser',
-            'teaser-selection/list@sulucontent',
+            'teaser-selection/list@suluarticle',
             [
                 'url' => '/admin/api/articles?locale={locale}',
                 'resultKey' => 'articles',
-                'searchFields' => ['title', 'type'],
-                'matchings' => [
-                    [
-                        'content' => 'public.title',
-                        'name' => 'title',
+                'searchFields' => ['title', 'route_path', 'changer_full_name', 'creator_full_name', 'author_full_name'],
+            ],
+            [
+                [
+                    'title' => $this->translator->trans('sulu_article.authored', [], 'backend'),
+                    'cssClass' => 'authored-slide',
+                    'contentSpacing' => true,
+                    'okDefaultText' => $okDefaultText,
+                    'buttons' => [
+                        [
+                            'type' => 'ok',
+                            'align' => 'right',
+                        ],
+                        [
+                            'type' => 'cancel',
+                            'align' => 'left',
+                        ],
                     ],
-                    [
-                        'content' => 'public.type',
-                        'name' => 'typeTranslation',
-                        'type' => 'translation',
+                ],
+                [
+                    'title' => $this->translator->trans('sulu_article.contact-selection-overlay.title', [], 'backend'),
+                    'cssClass' => 'contact-slide',
+                    'contentSpacing' => true,
+                    'okDefaultText' => $okDefaultText,
+                    'buttons' => [
+                        [
+                            'type' => 'ok',
+                            'align' => 'right',
+                        ],
+                        [
+                            'type' => 'cancel',
+                            'align' => 'left',
+                        ],
+                    ],
+                ],
+                [
+                    'title' => $this->translator->trans('sulu_article.category-selection-overlay.title', [], 'backend'),
+                    'cssClass' => 'category-slide',
+                    'contentSpacing' => true,
+                    'okDefaultText' => $okDefaultText,
+                    'buttons' => [
+                        [
+                            'type' => 'ok',
+                            'align' => 'right',
+                        ],
+                        [
+                            'type' => 'cancel',
+                            'align' => 'left',
+                        ],
+                    ],
+                ],
+                [
+                    'title' => $this->translator->trans('sulu_article.tag-selection-overlay.title', [], 'backend'),
+                    'cssClass' => 'tag-slide',
+                    'contentSpacing' => true,
+                    'okDefaultText' => $okDefaultText,
+                    'buttons' => [
+                        [
+                            'type' => 'ok',
+                            'align' => 'right',
+                        ],
+                        [
+                            'type' => 'cancel',
+                            'align' => 'left',
+                        ],
+                    ],
+                ],
+                [
+                    'title' => $this->translator->trans('public.choose', [], 'backend'),
+                    'cssClass' => 'page-slide data-source-slide',
+                    'contentSpacing' => false,
+                    'okDefaultText' => $okDefaultText,
+                    'buttons' => [
+                        [
+                            'type' => 'ok',
+                            'align' => 'right',
+                        ],
+                        [
+                            'type' => 'cancel',
+                            'align' => 'left',
+                        ],
                     ],
                 ],
             ]
@@ -99,13 +181,25 @@ class ArticleTeaserProvider implements TeaserProviderInterface
                 $excerpt->more,
                 $item->getRoutePath(),
                 count($excerpt->images) ? $excerpt->images[0]->id : $item->getTeaserMediaId(),
-                [
-                    'structureType' => $item->getStructureType(),
-                    'type' => $item->getType(),
-                ]
+                $this->getAttributes($item)
             );
         }
 
         return $result;
+    }
+
+    /**
+     * Returns attributes for teaser.
+     *
+     * @param ArticleViewDocument $viewDocument
+     *
+     * @return array
+     */
+    protected function getAttributes(ArticleViewDocument $viewDocument)
+    {
+        return [
+            'structureType' => $viewDocument->getStructureType(),
+            'type' => $viewDocument->getType(),
+        ];
     }
 }
